Keychain label support
The cosmetic difference is that generic passwords now have a label and GenericPassword.__repr__ will use it.

This involved adding or fixing several struct definitions, multiple constants and dealing with several misleading parts of the Keychain documentation. The next step will be to move this into the classes and expand the support beyond labels to allow any keychain attribute to be retrieved or, finally, updated.

Signed-off-by: Clinton Blackmore <clinton.blackmore@gmail.com>